% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn.R
\name{cdn}
\alias{cdn}
\title{cutdist/nnout table for hierarchical clustering method}
\usage{
cdn(x= NULL, cd=c(0.1, 0.5, 0.05), n=c(1, 5), spp_list=NULL, 
cutout=1, method="average", mdsmethod="classical")
}
\arguments{
\item{x}{A \emph{prab} object obtained from \code{prabclus::prabinit}.}

\item{cd}{A vector containing three numerical objects to be passed to \code{seq(from, to, by)} as: 1. (from) smaller cutdist value, 2. (to) larger cutdist value, and 3. increment of the sequence.}

\item{n}{A vector containing two numerical objects to be passed to \code{seq(from,to)} as: 1. (from) smaller nnout value, 2. (to) larger nnout value.}

\item{spp_list}{A data frame with one column containg the species names present in the \emph{prab} object.}

\item{cutout}{See \emph{cutout} in \code{?prabclus::hprabclust}.}

\item{method}{See \emph{method} in \code{?prabclus::hprabclust}.}

\item{mdsmethod}{See \emph{mdsmethod} in \code{?prabclus::hprabclust}.}
}
\value{
Returns a data.frame with the number of Biotic Elements and number of species assigned to the noise component for every cutdist/nnout combination given by the sequences created from the values provided in \emph{cd} and \emph{n}.
}
\description{
\code{cdn()} has been developed to improve the process of analysing different combinations of \emph{cutdist} and \emph{nnout} values passed to \code{prabclus::hprabclust()}, the hierarchical clustering method applied in Biotic Element Analyses
}
